Add reusable Cherry Pick by Comment workflow
Adds a new workflow YAML that can be used by repos to allow
pull request changes to be cherry-picked to another branch after
the pull request is merged.

This is done by leaving a comment in the PR after it is merged in
the format:
  `/auto-cherry-pick <target branch>`

For example, if a change is made on a `release/202208` branch, it
can be cherry-picked to `release/202202` by leaving a comment in
the PR after it is merged with:
  `/auto-cherry-pick release/202202`

The reusable workflow here should be used by repos that pick this
change up to centralize configuration of the GitHub Action.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>