Add Release Draft workflow
Adds a new workflow and accompanying configuration files that allow
releases to automatically be drafted in repos as pull requests are
completed.

Semantic versioning is automatically determined based on the standard
set of Project Mu labels associated with the pull request.

The types of changes (e.g. bug fix, new feature, etc.) are
automatically categorized in the pull request based on the labels as
well.

This makes tracking the new Semantic Version based on the type of
changes automatic and a release is always ready. In between releases,
the draft is a nice way to see a high-level delta since the last
release.

- `.github/ReleaseDrafter.yml` - Single reusable workflow to maintain
  a release draft as pull requests complete.
- `.github/ReleaseDraft.yml` - Configuration file for the
  ReleaseDrafter workflow. Defines label meanings for releases and
  how releases should be drafted.
- `.sync/workflows/leaf/release-draft.yml` - A leaf workflow that can
  be synced to repos to trigger the reusable workflow in Mu DevOps.

  Project Mu repos can easily opt into consistent release drafting by
  including this leaf workflow via file sync.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>