.sync/dependabot: Do not autorebase pull requests
Dependabot is currently enabled in at least 13 Project Mu repos. The
default `rebase-strategy` is auto which means dependabot will rebase
open pull requests when changes are detected.

Most often this occurs because another pull request was merged and
dependabot recognizes its pull request is now in conflict with that
merged change (e.g. another dependabot change).

The impact of this is that a lot of CI time is spent building
dependabot pull requests that are not actually going to be merged
because they do not have the proper number of reviews.

This change sets the strategy to auto. Once we recognize that a
dependabot PR has the number of reviews needed, we manually rebase
it in the GitHub UI and set auto-merge.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>