Add N, N-1, and main branch release drafting
Updates the `ReleaseDrafter.yml` workflow to filter the base ref to
use when composing the release draft based on the GitHub trigger
context. This allows simultaneous release drafting in repos that
use release branches while repos with non-release branches (e.g.
main branch) are not impacted.

Adds a variable to `Version.njk` to track the previous Mu release
branch name. This files continues to serve as a single location
where the release branches are defined.

Updates the file sync process to sync the release drafter config
file as needed for the corresponding repo.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>