Dockerfile: Add Rust support
Updates the `build` container image to be able to build Rust code
(including Rust UEFI code).

The following additional applications are installed to assist with
setting up Rust dependencies:

  `curl`, `jq`, and `unzip`

The `CARGO_HOME` and `RUSTUP_HOME` environment variables are set in
the image and the cargo directory is added to the system path.

The version of Rust specified in `Version.njk` is installed alongside
support for other tools used in Project Mu Rust builds such as
`cargo-make` and `cargo-tarpaulin`. The latest release binaries are
pulled directly from their GitHub releases to minimze container
size impact.

The resulting container images can be used for local development and
within pipelines.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>