steps/RustSetupSteps.yml: Use step template to download artifacts
Downloads the Cargo tools from an Azure pipeline using a YAML step
template. The Python code is inline so it can directly be used as a
template in other YAML files that are used as a repository resource.
If in a separate Python file, the mu_devops repo would need to be
checked out to use access the file.

Signed-off-by: Michael Kubacki <michael.kubacki@microsoft.com>